+++ scripts/doc.sh
@@ -4,7 +4,7 @@
 mkdir -p doc/ref
 
 # react-docgen
-react-docgen src/Storage.js src/Source.js src/Pinnable.js src/Balloon.js src/Cloud.js \
+npx react-docgen src/Storage.js src/Source.js src/Pinnable.js src/Balloon.js src/Cloud.js \
  -o stories/docgen.json  --resolver=findAllComponentDefinitions
 
 # react-docgen to md